+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changes
    ([#issue/31](http://github.com/RubyMoney/money/issues/31))
  - Removed deprecated `Money::SYMBOLS`, `Money::SEPARATORS` and `Money::DELIMITERS`.
  - Removed deprecated `Money::VariableExchangeBank`.
+ - Removed deprecated `Money::Bank::Base#exchange`.
 
 Money 3.1.0
 ===========